test: 큐랑 리스트 테스트하는 코드를 각각 다른 파일로 분리
