+++ Calculator/Calculator/Node/LinkedList.swift
@@ -1,14 +1,3 @@
-//class Node<T> {
-//    var data: T
-//    var next: Node?
-//
-//    init(data: T, next: Node? = nil) {
-//        self.data = data
-//        self.next = next
-//    }
-//}
-
-
 class LinkedList<T> {
     var head: Node<T>?
     var tail: Node<T>?
